requirements.txt
Django==1.5.2
path.py==4.3
requests==1.2.3